+++ spirit-sensor-gateway/domain/SpiritTrack.h
@@ -23,19 +23,19 @@ class SpiritTrack {
         ~SpiritTrack();
         uint16_t getTrackID();
         void setDistance(uint16_t distance);
-        void setSpeed(uint16_t speed);
-        void setAcceleration(uint16_t acceleration);
+        void setSpeed(int16_t speed);
+        void setAcceleration(int16_t acceleration);
         uint16_t getDistance();
-        uint16_t getSpeed();
-        uint16_t getAcceleration();
+        int16_t getSpeed();
+        int16_t getAcceleration();
         uint16_t getIntensity();
-        uint16_t getConfidenceLevel();
+        uint8_t getConfidenceLevel();
 
     private:
         uint16_t id;
         uint16_t distance = 0;
-        uint16_t speed = 0;
-        uint16_t acceleration = 0;
+        int16_t speed = 0;
+        int16_t acceleration = 0;
         uint16_t intensity;
         uint8_t confidenceLevel;
 };
